% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_two_datasets.R
\name{test_two_datasets}
\alias{test_two_datasets}
\title{Regression testing for dataset migration}
\usage{
test_two_datasets(legacy_file = NULL, legacy_df = NULL,
  target_file = NULL, target_df = NULL, hier = NULL,
  hier_df = NULL, hier_pair = NULL, hier_pair_df = NULL,
  thresholds = NULL, thresholds_df = NULL, final_report = NULL,
  final_data = NULL, key_col, hier_col, report_join = TRUE,
  report_hybrid = TRUE, report_magnitude = TRUE, report_mre = TRUE,
  report_spearman = TRUE, report_pearson = TRUE,
  report_distribution = TRUE, report_rank = TRUE,
  report_spearman_diff = TRUE, report_na = TRUE,
  report_var_attr = TRUE)
}
\arguments{
\item{legacy_file}{Full path of the input legacy dataset (csv)}

\item{legacy_df}{Data frame contained the input legacy dataset}

\item{target_file}{Full path of the input target dataset (csv)}

\item{target_df}{Data frame contained the input target dataset}

\item{hier}{Full path of the hierarchy file (csv)}

\item{hier_df}{Data frame contained the hierarchy}

\item{hier_pair}{Full path of the hierarchical pair file (csv)}

\item{hier_pair_df}{Data frame contained the hierarchical pair}

\item{thresholds}{Full path of the file contained thresholds}

\item{thresholds_df}{Data frame contained thresholds}

\item{final_report}{Full path of the output file (xlsx)}

\item{final_data}{Full path of the output file (RData)}

\item{key_col}{Key column in the two datasets}

\item{hier_col}{Column name that contains hierarchies}

\item{report_join}{Boolean variable to control the report of joined metrics. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_hybrid}{Boolean variable to control the report of hybrid metrics. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_magnitude}{Boolean variable to control the report of magnitude metrics. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_mre}{Boolean variable to control the report based on mean relative errors. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_spearman}{Boolean variable to control the report based on Spearman's test. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_pearson}{Boolean variable to control the report based on Pearson test. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_distribution}{Boolean variable to control the report based on distribution test. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_rank}{Boolean variable to control the report of appearances. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_spearman_diff}{Boolean variable to control the report of Spearman's test difference on different hierarchies. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_na}{Boolean variable to control the report of features with NA values. TRUE - generate the report; FALSE - the report will not be generated.}

\item{report_var_attr}{Boolean variable to control the report of variables' attributes. TRUE - generate the report; FALSE - the report will not be generated.}
}
\description{
This is a package of regression testing for dataset migration. Given a scenario where a legacy dataset will be replaced by a target dataset, we will analyze the difference between them based on following tests:
1. Distribution test: Kolmogorov-Smirnov test.
2. Correlation tests: Pearson correlation coefficient and Spearman's correlation.
3. Mean relative errors.
4. The difference between two hierarchical pairs in Spearman's test.
5. Features that have NA values.
6. Hybrid tests, which shows features that appear in Kolmogorov-Smirnov test, mean relative error test, and correlation tests.
The final report will be written into a user-specified xlsx file or an object (which is stored in an RData file). Users can choose the test results produced in the final report.
}
